update vim/ycm installation for 24.04

+++ appconfig/vim/install.sh
@@ -42,7 +42,9 @@ while true; do
 
     toilet Setting up vim
 
-    sudo apt-get -y remove vim-* || echo ""
+    if [ -z "$NUMBAT" ]; then
+      sudo apt-get -y remove vim-* || echo ""
+    fi
 
     if [ -n "$BEAVER" ]; then
       sudo apt-get -y install libgnome2-dev libgnomeui-dev libbonoboui2-dev
@@ -56,16 +58,19 @@ while true; do
       sudo -H pip3 install rospkg
     fi
 
-    # compile vim from sources
-    cd $APP_PATH/../../submodules/vim
-    ./configure --with-features=huge \
-      --enable-multibyte \
-      --enable-python3interp=yes \
-      --with-python3-config-dir=/usr/lib/python3.5/config-3.5m-x86_64-linux-gnu \
-      --enable-perlinterp=yes \
-      --enable-luainterp=yes \
-      --enable-gui=no \
-      --enable-cscope --prefix=/usr
+    if [ -n "$NUMBAT" ]; then
+      sudo apt-get -y install vim-gtk3
+    else
+      # compile vim from sources
+      cd $APP_PATH/../../submodules/vim
+      ./configure --with-features=huge \
+        --enable-multibyte \
+        --enable-python3interp=yes \
+        --with-python3-config-dir=/usr/lib/python3.5/config-3.5m-x86_64-linux-gnu \
+        --enable-perlinterp=yes \
+        --enable-luainterp=yes \
+        --enable-gui=no \
+        --enable-cscope --prefix=/usr
 
       ## add for python2
       # --enable-pythoninterp=yes \
@@ -80,6 +85,7 @@ while true; do
       cd ../
       make VIMRUNTIMEDIR=/usr/share/vim/vim81
       sudo make install
+    fi
 
     # set vim as a default git mergetool
     git config --global merge.tool vimdiff
@@ -91,67 +97,69 @@ while true; do
     # updated new plugins and clean old plugins
     /usr/bin/vim -E -c "let g:user_mode=1" -c "so $APP_PATH/dotvimrc" -c "PlugInstall" -c "wqa" || echo "It normally returns >0"
 
-    default=y
-    while true; do
-      if [[ "$unattended" == "1" ]]
-      then
-        resp=$default
-      else
-        [[ -t 0 ]] && { read -t 10 -n 2 -p $'\e[1;32mCompile YouCompleteMe? [y/n] (default: '"$default"$')\e[0m\n' resp || resp=$default ; }
-      fi
-      response=`echo $resp | sed -r 's/(.*)$/\1=/'`
-
-      if [[ $response =~ ^(y|Y)=$ ]]
-      then
-
-        # set youcompleteme
-        toilet Setting up youcompleteme
-
-        # if not on 20.04, g++-8 has to be installed manually
-        if [ -n "$BEAVER" ]; then
-          sudo apt-get -y install g++-8
-          sudo update-alternatives --install /usr/bin/gcc gcc /usr/bin/gcc-7 700 --slave /usr/bin/g++ g++ /usr/bin/g++-7
-          sudo update-alternatives --install /usr/bin/gcc gcc /usr/bin/gcc-8 800 --slave /usr/bin/g++ g++ /usr/bin/g++-8
-          # add llvm repo for clangd and python3-clang
-          wget -O - https://apt.llvm.org/llvm-snapshot.gpg.key | sudo apt-key add -
-          sudo apt-add-repository "deb http://apt.llvm.org/bionic/ llvm-toolchain-bionic-11 main"
-          # if 18.04, python3-clang has to be installed throught pip3 with prequisites manually from apt
-          sudo apt-get -y install clang-11 libclang-11-dev
-          sudo pip3 install clang
+    if [-z "$NUMBAT" ]; then
+      default=y
+      while true; do
+        if [[ "$unattended" == "1" ]]
+        then
+          resp=$default
         else
-          # if 20.04, just install python3-clang from apt
-          sudo apt-get -y install python3-clang 
+          [[ -t 0 ]] && { read -t 10 -n 2 -p $'\e[1;32mCompile YouCompleteMe? [y/n] (default: '"$default"$')\e[0m\n' resp || resp=$default ; }
         fi
+        response=`echo $resp | sed -r 's/(.*)$/\1=/'`
 
-        if [ -n "$NUMBAT" ]; then
-          # install prequisites for YCM
-          sudo apt-get -y install clangd-18
-          # set clangd to version 18 by default
-          sudo update-alternatives --install /usr/bin/clangd clangd /usr/bin/clangd-18 999
-          sudo apt-get -y install libboost-all-dev
+        if [[ $response =~ ^(y|Y)=$ ]]
+        then
+
+          # set youcompleteme
+          toilet Setting up youcompleteme
+
+          # if not on 20.04, g++-8 has to be installed manually
+          if [ -n "$BEAVER" ]; then
+            sudo apt-get -y install g++-8
+            sudo update-alternatives --install /usr/bin/gcc gcc /usr/bin/gcc-7 700 --slave /usr/bin/g++ g++ /usr/bin/g++-7
+            sudo update-alternatives --install /usr/bin/gcc gcc /usr/bin/gcc-8 800 --slave /usr/bin/g++ g++ /usr/bin/g++-8
+            # add llvm repo for clangd and python3-clang
+            wget -O - https://apt.llvm.org/llvm-snapshot.gpg.key | sudo apt-key add -
+            sudo apt-add-repository "deb http://apt.llvm.org/bionic/ llvm-toolchain-bionic-11 main"
+            # if 18.04, python3-clang has to be installed throught pip3 with prequisites manually from apt
+            sudo apt-get -y install clang-11 libclang-11-dev
+            sudo pip3 install clang
+          else
+            # if 20.04, just install python3-clang from apt
+            sudo apt-get -y install python3-clang 
+          fi
+
+          if [ -n "$NUMBAT" ]; then
+            # install prequisites for YCM
+            sudo apt-get -y install clangd-18
+            # set clangd to version 18 by default
+            sudo update-alternatives --install /usr/bin/clangd clangd /usr/bin/clangd-18 999
+            sudo apt-get -y install libboost-all-dev
+          else
+            # install prequisites for YCM
+            sudo apt-get -y install clangd-11
+            # set clangd to version 11 by default
+            sudo update-alternatives --install /usr/bin/clangd clangd /usr/bin/clangd-11 999
+            sudo apt-get -y install libboost-all-dev
+          fi
+
+          cd ~/.vim/plugged/YouCompleteMe/
+          git submodule update --init --recursive
+          python3 ./install.py --clangd-completer
+
+          # link .ycm_extra_conf.py
+          ln -fs $APP_PATH/dotycm_extra_conf.py ~/.ycm_extra_conf.py
+
+          break
+        elif [[ $response =~ ^(n|N)=$ ]]
+        then
+          break
         else
-          # install prequisites for YCM
-          sudo apt-get -y install clangd-11
-          # set clangd to version 11 by default
-          sudo update-alternatives --install /usr/bin/clangd clangd /usr/bin/clangd-11 999
-          sudo apt-get -y install libboost-all-dev
+          echo " What? \"$resp\" is not a correct answer. Try y+Enter."
         fi
-
-        cd ~/.vim/plugged/YouCompleteMe/
-        git submodule update --init --recursive
-        python3 ./install.py --clangd-completer
-
-        # link .ycm_extra_conf.py
-        ln -fs $APP_PATH/dotycm_extra_conf.py ~/.ycm_extra_conf.py
-
-        break
-      elif [[ $response =~ ^(n|N)=$ ]]
-      then
-        break
-      else
-        echo " What? \"$resp\" is not a correct answer. Try y+Enter."
-      fi
-    done
+      done
+    fi
 
     break
   elif [[ $response =~ ^(n|N)=$ ]]
